Phase 12 achieved 9 error fixes (3.7% reduction) bringing total to 43.6% improvement.

Key fixes:
- union-attr (3): Fixed regression from Phase 11.1
- return-value (1): Float return type for _maze_distance
- misc (1): Removed ClassVar from metadata override
- operator+arg-type (4): Added explicit type annotations

Result: 246 → 237 errors
Cumulative: 423 → 237 errors (43.6% improvement)

+++ mfg_pde/alg/reinforcement/environments/position_placement.py
@@ -180,7 +180,7 @@ def _place_clustered(grid: Grid, num_positions: int) -> list[tuple[int, int]]:
         return [center, *neighbors, *remaining]
 
 
-def _maze_distance(grid: Grid, start: tuple[int, int], goal: tuple[int, int]) -> int:
+def _maze_distance(grid: Grid, start: tuple[int, int], goal: tuple[int, int]) -> float:
     """
     Compute maze distance between two positions using BFS.
